+++ CHANGELOG.md
@@ -1,6 +1,7 @@
+* Compatibility: Thread-safe patch for ReGrowth 2.
 * Fix: Exception during game launch when both YaOpt and Performance Fish are installed, but Performance Fish's TryFindBestIngredientsHelpers_InnerDelegate is disabled.
 * Fix: System.InvalidOperationException when both Target Finding Optimization and Constant Job Prediction are enabled.
-* Fix: Thread-safe for ListerBuildingsRepairable.
+* Fix: Thread-safe for ListerBuildingsRepairable and MineAIUtility.
 
 ## 1.1.0